refactor: 서비스 인터페이스 제거 & 트랜잭션 메서드 단위
